refactor(rpc): extract aggregate executor helpers
Move server-side aggregate grouping, merging, and result emission helpers out of lineairdb_rpc.cc into aggregate_executor.{cc,hh}.
Register aggregate_executor.cc in the server build.
